+++ lib/netzke/core.rb
@@ -41,7 +41,7 @@ module Netzke
     @@external_ext_css = []
 
     mattr_accessor :icons_uri
-    @@icons_uri = "/images/icons"
+    @@icons_uri = "/assets/icons"
 
     mattr_accessor :ext_uri
     @@ext_uri = "/extjs"