+++ spec/rails_app/app/controllers/application_controller 2.rb
@@ -1,10 +1,10 @@
 class ApplicationController < ActionController::Base
+  # protect_from_forgery
+
   before_filter :set_locale
   def set_locale
     # if params[:locale] is nil then I18n.default_locale will be used
     I18n.locale = params[:locale]
   end
-  # Prevent CSRF attacks by raising an exception.
-  # For APIs, you may want to use :null_session instead.
-  protect_from_forgery with: :exception
+
 end