+++ spec/rails_app/app/controllers/simple_rails_controller.rb
@@ -1,2 +1,0 @@
-class SimpleRailsController < ApplicationController
-end