+++ spec/rails_app/config/initializers/secret_token.rb
@@ -1,7 +1,12 @@
 # Be sure to restart your server when you modify this file.
 
-# Your secret key for verifying the integrity of signed cookies.
+# Your secret key is used for verifying the integrity of signed cookies.
 # If you change this key, all old signed cookies will become invalid!
+
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-RailsApp::Application.config.secret_token = '37d8da5a1af6fb203bcbdb7ad5a944a8ed0fdc568594b9000319294a8b696d607dacf64bc521bdf7036790f7e1a65627bab7df5a29cd58d78239b7e6dd8377ae'
+# You can use `rake secret` to generate a secure secret key.
+
+# Make sure your secret_key_base is kept private
+# if you're sharing your code publicly.
+RailsApp::Application.config.secret_key_base = 'bbd2ca5b4336b1112a4b12f92ec13f6434e6b504f7582912a2903bae8fe0547303f86f5b910205264fa8d37dfc7206a63ab65f1c20a6e48e837d9c34b78f1408'